perbaikan session

+++ tampil_barang.php
@@ -7,9 +7,20 @@ if (!isset($_SESSION["nama"])) {
 }
 if ($_SESSION['level'] == 2 && $_SESSION['status'] == 1) {
     // Login berhasil dengan level staff dan status aktif, arahkan ke halaman lain
-    header("location: tampil_transaksi.php");
+    header("location: index.php");
     exit(); // Berhenti eksekusi skrip setelah mengalihkan header
 }
+
+// $level = $_SESSION['level'];
+// $edit_hapus_liat = false;
+// if ($level == 1) {
+//     $edit_hapus_liat = true;
+// } elseif ($level == 4) {
+//     $edit_hapus_liat = true;
+// } elseif ($level == 3) {
+//     $edit_hapus_liat = true;
+// }
+
 ?>
 <!DOCTYPE html>
 <html lang="en">
@@ -54,8 +65,10 @@ if ($_SESSION['level'] == 2 && $_SESSION['status'] == 1) {
                 <td><?php echo $d['harga']; ?></td>
                 <td><?php echo $d['nama_kategori']; ?></td>
                 <td>
-                    <a href="edit_barang.php?id=<?php echo $d['id']; ?>">EDIT</a> |
-                    <a href="hapus_barang.php?id=<?php echo $d['id']; ?>">HAPUS</a>
+                    <?php if ($level != 2) : ?>
+                        <a href="edit_barang.php?id=<?php echo $d['id']; ?>">EDIT</a> |
+                        <a href="hapus_barang.php?id=<?php echo $d['id']; ?>">HAPUS</a>
+                    <?php endif; ?>
                 </td>
             </tr>
         <?php
